+++ CHANGELOG.md
@@ -10,6 +10,10 @@ use [Keep a Changelog] (http://keepachangelog.com/).
 
 - Suggest `psr/log` dependency, required to use the `LogAdapter`
 
+### Fixed
+
+- `FallbackAdapter` no longer throws `null` and returns the last result (even `false`) when every adapter answered, instead of re-throwing an exception raised by the last adapter
+
 ## [1.0.1] - 2024-03-20
 
 ### Fixed